refactor: renamed package from example to demal
