+++ feat-profile/src/main/java/com/demal/feat_profile/navigation/ProfileNavigator.kt
@@ -1,4 +1,4 @@
-package com.example.feat_profile.navigation
+package com.demal.feat_profile.navigation
 
 interface ProfileNavigator {
     //Методы для навигации ИЗ ProfileFragment
